move MessageUUID to wrapper-common module
