+++ sphinx/application/data/features/feature-repository/src/main/java/chat/sphinx/feature_repository/model/MessageDboWrapper.kt
@@ -10,6 +10,7 @@ import chat.sphinx.wrapper_common.lightning.LightningPaymentHash
 import chat.sphinx.wrapper_common.lightning.LightningPaymentRequest
 import chat.sphinx.wrapper_common.lightning.Sat
 import chat.sphinx.wrapper_common.message.MessageId
+import chat.sphinx.wrapper_common.message.MessageUUID
 import chat.sphinx.wrapper_message.*
 import chat.sphinx.wrapper_message.media.MessageMedia
